requirements-base.txt
pytest
coverage
flake8
black
isort
pytest-cov
codecov
mypy
gitchangelog
mkdocs
openai
langchain
chainlit
pinecone-client
llama_index
transformers
ndg-httpsclient
pyopenssl
pyasn1
waiting
python-dotenv
firebase-admin==4.4.0
requests==2.31.0
urllib3==1.26.15
Flask-Cors==3.0.8
google-auth-oauthlib==1.0.0